+++ android/src/main/java/expo/modules/yandexmapkit/ExpoYandexMapKitModule.kt
@@ -110,6 +110,26 @@ class ExpoYandexMapKitModule : Module() {
       OnViewDidUpdateProps { view: ExpoYandexMapKitView ->
         view.applyPendingCameraPosition()
       }
+
+      AsyncFunction("setCenter") { view: ExpoYandexMapKitView, position: CameraPositionRecord, options: CameraMoveOptionsRecord ->
+        view.moveCamera(position, options)
+      }
+
+      AsyncFunction("getCameraPosition") { view: ExpoYandexMapKitView ->
+        view.currentCameraPosition()
+      }
+
+      AsyncFunction("getVisibleRegion") { view: ExpoYandexMapKitView ->
+        view.currentVisibleRegion()
+      }
+
+      AsyncFunction("getScreenPoints") { view: ExpoYandexMapKitView, points: List<PointRecord> ->
+        view.screenPoints(points)
+      }
+
+      AsyncFunction("getWorldPoints") { view: ExpoYandexMapKitView, points: List<ScreenPointRecord> ->
+        view.worldPoints(points)
+      }
     }
   }